# optimize performance and minor refactor